[change] move app to its own module
