+++ ata_api/main.py
@@ -4,32 +4,17 @@ from typing import Annotated
 from uuid import UUID
 
 from ata_db_models.models import Group
-from fastapi import Depends, FastAPI, Path, Query, Request, status
-from fastapi.responses import JSONResponse
+from fastapi import Depends, Path, Query
 from mangum import Mangum
 from sqlalchemy.orm import Session
-from starlette.exceptions import ExceptionMiddleware
 
+from ata_api.app import app
 from ata_api.crud import create_prescription, read_prescription
 from ata_api.db import create_db_session
 from ata_api.models import PrescriptionResponse
 from ata_api.monitoring.logging import logger
 from ata_api.monitoring.metrics import metrics
-from ata_api.router import LoggerRouteHandler
 from ata_api.site import SiteName
-
-app = FastAPI()
-# Add FastAPI context to logs
-app.router.route_class = LoggerRouteHandler
-# Add exception middleware to log unhandled exceptions
-app.add_middleware(ExceptionMiddleware, handlers=app.exception_handlers)
-
-
-@app.exception_handler(Exception)
-async def unhandled_exception_handler(request: Request, exception: Exception) -> JSONResponse:
-    logger.exception("Unhandled exception")
-    return JSONResponse(status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, content={"detail": "Internal Server Error"})
-
 
 logger.info(f"AtA API starting up on stage {os.environ.get('STAGE')}")
